fixup: try_cast: move test
